Reformat extract class validateTask method logic
